Revert to March 2015 data defaults as the newer one seems messy - for example missing Abbey View Medical Centre (epraccur says 'Dormant', SYOA does not have info).
Added checking to see where (epraccur? SYOA? Dorset localities?) there may be missing information.  *** This is done using a merge(..., all = TRUE) - and therefore CCGcodes.DorsetAndSurroundings now returns 676 GP surgeries (not 675 as was used in the CSR2015 calculation, however the extra one was a SouthGloucestershire one, so not critical to the core calculations in that document)